refactor: phase 1 service split — macro, strategy, api.ts (partial)
## macro_market_service.py → 2 files + shim
- api/services/macro_service.py (975L): MacroMarketService(MarketDataMixin)
  — signal computation, history, DB persistence, backfill
- api/services/market_data_service.py (453L): MarketDataMixin
  — real-time fetch adapters (FX, futures, flow, breadth, Naver)
- api/services/macro_market_service.py → 12-line backward-compat shim

## strategy_service.py → 5 files + shims
- api/services/strategy/strategy_core_service.py (430L): condition registry,
  graph traversal
- api/services/strategy/strategy_execution_service.py (355L): universe
  resolution, screening, execution
- api/services/strategy/strategy_analytics_service.py (316L): backtest
  results, comparison, leaderboard
- api/services/strategy/strategy_fundamentals.py (216L): PER/PBR enrichment
- api/services/strategy/strategy_portfolio_helpers.py (502L): weight math,
  ranking, node survivors
- strategy_service.py / strategy_comparison_service.py /
  strategy_backtest_result_service.py → thin shims (all existing imports intact)

## web/src/lib/api.ts (1885L) → 9 domain files
- api/_base.ts: fetchApi + API_BASE_URL
- api/screeningApi.ts, portfolioApi.ts, analysisApi.ts, marketApi.ts,
  strategyApi.ts, brokerApi.ts, reportApi.ts
- api/index.ts: barrel re-export (all `@/lib/api` imports unchanged)
- TypeScript strict: 0 errors

## Docs
- docs/works/20260427_directory_restructure.md: full restructure plan
