+++ web/src/lib/api/index.ts
@@ -1,0 +1,11 @@
+// Re-export the base fetch utility so consumers of '@/lib/api' can still
+// import fetchApi from this barrel entry-point.
+export { fetchApi, API_BASE_URL } from './_base';
+
+export * from './screeningApi';
+export * from './portfolioApi';
+export * from './analysisApi';
+export * from './marketApi';
+export * from './strategyApi';
+export * from './brokerApi';
+export * from './reportApi';
